move instance tests
