+++ modules/core/arrow-core/build.gradle
@@ -8,6 +8,7 @@ dependencies {
     testCompile "io.kotlintest:kotlintest:$kotlinTestVersion"
     testCompile project(':arrow-test')
     testCompile project(':arrow-syntax')
+    testCompile project(':arrow-instances-core')
 }
 
 apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
