+++ modules/core/arrow-core/src/test/kotlin/arrow/instances/NumberMonoidTest.kt
@@ -1,6 +1,5 @@
-package arrow.typeclasses
+package arrow.instances
 
-import arrow.instances.*
 import arrow.test.UnitSpec
 import io.kotlintest.KTestJUnitRunner
 import io.kotlintest.properties.forAll
